Correct typo in docstring (int -> float)
This is based on https://stackoverflow.com/q/71494698/10693596

+++ networkx/classes/function.py
@@ -1288,7 +1288,7 @@ def path_weight(G, path, weight):
 
     Returns
     -------
-    cost: int
+    cost: float
         A integer representing the total cost with respect to the
         specified weight of the specified path
